ShaderChunk: Rename uvs_* to uv_* back.
