+++ examples/webgl_gpgpu_water.html
@@ -186,7 +186,7 @@
 			#endif
 
 			#include <common>
-			#include <uvs_pars_vertex>
+			#include <uv_pars_vertex>
 			#include <displacementmap_pars_vertex>
 			#include <envmap_pars_vertex>
 			#include <color_pars_vertex>
@@ -200,7 +200,7 @@
 
 				vec2 cellSize = vec2( 1.0 / WIDTH, 1.0 / WIDTH );
 
-				#include <uvs_vertex>
+				#include <uv_vertex>
 				#include <color_vertex>
 
 				// # include <beginnormal_vertex>
