+++ manual/en/indexed-textures.html
@@ -293,7 +293,7 @@ we can modify the default shaders by adding a function to a material's
 <p>The default fragment shader looks something like this before compiling.</p>
 <pre class="prettyprint showlinemods notranslate lang-glsl" translate="no">#include &lt;common&gt;
 #include &lt;color_pars_fragment&gt;
-#include &lt;uvs_pars_fragment&gt;
+#include &lt;uv_pars_fragment&gt;
 #include &lt;map_pars_fragment&gt;
 #include &lt;alphamap_pars_fragment&gt;
 #include &lt;aomap_pars_fragment&gt;
