+++ manual/ja/indexed-textures.html
@@ -273,7 +273,7 @@ three.jsでデフォルトのシェーダーを変更してみましょう。
 <p>デフォルトのフラグメントシェーダーはコンパイル前は以下のようになっています。</p>
 <pre class="prettyprint showlinemods notranslate lang-glsl" translate="no">#include &lt;common&gt;
 #include &lt;color_pars_fragment&gt;
-#include &lt;uvs_pars_fragment&gt;
+#include &lt;uv_pars_fragment&gt;
 #include &lt;map_pars_fragment&gt;
 #include &lt;alphamap_pars_fragment&gt;
 #include &lt;aomap_pars_fragment&gt;
