+++ manual/ko/indexed-textures.html
@@ -240,7 +240,7 @@ const palette = [
 <p>아래는 내장 fragment 쉐이더를 수정하기 전입니다.</p>
 <pre class="prettyprint showlinemods notranslate lang-glsl" translate="no">#include &lt;common&gt;
 #include &lt;color_pars_fragment&gt;
-#include &lt;uvs_pars_fragment&gt;
+#include &lt;uv_pars_fragment&gt;
 #include &lt;map_pars_fragment&gt;
 #include &lt;alphamap_pars_fragment&gt;
 #include &lt;aomap_pars_fragment&gt;
