+++ manual/zh/indexed-textures.html
@@ -264,7 +264,7 @@ const palette = [
         <p>默认的片元着色器在编译之前看起来就像这样：</p>
         <pre class="prettyprint showlinemods notranslate lang-glsl" translate="no">#include &lt;common&gt;
 #include &lt;color_pars_fragment&gt;
-#include &lt;uvs_pars_fragment&gt;
+#include &lt;uv_pars_fragment&gt;
 #include &lt;map_pars_fragment&gt;
 #include &lt;alphamap_pars_fragment&gt;
 #include &lt;aomap_pars_fragment&gt;
