+++ src/renderers/shaders/ShaderLib/depth.glsl.js
@@ -1,6 +1,6 @@
 export const vertex = /* glsl */`
 #include <common>
-#include <uvs_pars_vertex>
+#include <uv_pars_vertex>
 #include <displacementmap_pars_vertex>
 #include <morphtarget_pars_vertex>
 #include <skinning_pars_vertex>
@@ -14,7 +14,7 @@ varying vec2 vHighPrecisionZW;
 
 void main() {
 
-	#include <uvs_vertex>
+	#include <uv_vertex>
 
 	#include <skinbase_vertex>
 
@@ -48,7 +48,7 @@ export const fragment = /* glsl */`
 
 #include <common>
 #include <packing>
-#include <uvs_pars_fragment>
+#include <uv_pars_fragment>
 #include <map_pars_fragment>
 #include <alphamap_pars_fragment>
 #include <alphatest_pars_fragment>
