+++ src/renderers/shaders/ShaderLib/linedashed.glsl.js
@@ -5,7 +5,7 @@ attribute float lineDistance;
 varying float vLineDistance;
 
 #include <common>
-#include <uvs_pars_vertex>
+#include <uv_pars_vertex>
 #include <color_pars_vertex>
 #include <fog_pars_vertex>
 #include <morphtarget_pars_vertex>
@@ -16,7 +16,7 @@ void main() {
 
 	vLineDistance = scale * lineDistance;
 
-	#include <uvs_vertex>
+	#include <uv_vertex>
 	#include <color_vertex>
 	#include <morphcolor_vertex>
 	#include <begin_vertex>
@@ -40,7 +40,7 @@ varying float vLineDistance;
 
 #include <common>
 #include <color_pars_fragment>
-#include <uvs_pars_fragment>
+#include <uv_pars_fragment>
 #include <map_pars_fragment>
 #include <fog_pars_fragment>
 #include <logdepthbuf_pars_fragment>
