+++ src/renderers/shaders/ShaderLib/meshbasic.glsl.js
@@ -1,6 +1,6 @@
 export const vertex = /* glsl */`
 #include <common>
-#include <uvs_pars_vertex>
+#include <uv_pars_vertex>
 #include <envmap_pars_vertex>
 #include <color_pars_vertex>
 #include <fog_pars_vertex>
@@ -11,7 +11,7 @@ export const vertex = /* glsl */`
 
 void main() {
 
-	#include <uvs_vertex>
+	#include <uv_vertex>
 	#include <color_vertex>
 	#include <morphcolor_vertex>
 
@@ -52,7 +52,7 @@ uniform float opacity;
 #include <common>
 #include <dithering_pars_fragment>
 #include <color_pars_fragment>
-#include <uvs_pars_fragment>
+#include <uv_pars_fragment>
 #include <map_pars_fragment>
 #include <alphamap_pars_fragment>
 #include <alphatest_pars_fragment>
