+++ src/renderers/shaders/ShaderLib/meshnormal.glsl.js
@@ -8,7 +8,7 @@ export const vertex = /* glsl */`
 #endif
 
 #include <common>
-#include <uvs_pars_vertex>
+#include <uv_pars_vertex>
 #include <displacementmap_pars_vertex>
 #include <normal_pars_vertex>
 #include <morphtarget_pars_vertex>
@@ -18,7 +18,7 @@ export const vertex = /* glsl */`
 
 void main() {
 
-	#include <uvs_vertex>
+	#include <uv_vertex>
 
 	#include <beginnormal_vertex>
 	#include <morphnormal_vertex>
@@ -56,7 +56,7 @@ uniform float opacity;
 #endif
 
 #include <packing>
-#include <uvs_pars_fragment>
+#include <uv_pars_fragment>
 #include <normal_pars_fragment>
 #include <bumpmap_pars_fragment>
 #include <normalmap_pars_fragment>
